+++ com.unity.render-pipelines.universal/CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Optimized 2D Renderer performance on mobile GPUs by reducing the number of render target switches.
 - Optimized 2D Renderer performance by rendering the normal buffer at the same lower resolution as the light buffers.
+- Improved Light2D UI/UX
+- Improved 2D Menu layout
+- Deprecated Light2D Parametric Light
+- Deprecated Light2D point light cookie
+- Renamed Light2D point light to spot light
+- Added preset shapes for creating a freeform light
+- Added 2D Renderer Asset Preset for creating a Universal Renderer Asset
 - 2D Renderer: The per Blend Style render texture scale setting was replaced by a global scale setting for all Blend Styles.
 - Optimized 2D Renderer performance by using a tiny light texture for layer/blend style pairs for which no light is rendered.
 - Reorgnized the settings in 2D Renderer Data Inspector.